requirements.txt
PyQt5 # For GUI.
pywin32 # For interacting with Powerpoint.
polars # For csv processing.
colorama # For console color.
validators # For URL validation.
gdown # For downloading files from Google Drive.
pystache # For text replacement.
pillow # For image processing.

# Add-ons
pillow-heif # HEIF Pillow supports.
pillow-avif-plugin # AVIF Pillow supports.